backend/requirements.txt
django
django-cors-headers
djangorestframework
markdown
psycopg2
django-oauth-toolkit
factory-boy
django-extensions
Pillow
django-storages
coverage
freezegun
black
isort